CI/code-quality-checks:     - Add linter and formatter, and integrate into CI pipeline.     - Run initial formatter pass.     - Added check for correct formatting in to linter.

+++ contracts/Factory.sol
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.3;
 
-import "./Synth.sol"; 
+import "./Synth.sol";
 
 // Factory Contract
 contract Factory {
-
     bool private inited;
     address public VADER;
     address public USDV;
@@ -21,9 +20,9 @@ contract Factory {
         require(msg.sender == POOLS, "!POOLS");
         _;
     }
-    
-    constructor(){
-    }
+
+    constructor() {}
+
     function init(address _pool) public {
         require(inited == false);
         inited = true;
@@ -31,17 +30,21 @@ contract Factory {
     }
 
     //Create a synth asset
-    function deploySynth(address token) external onlyPOOLS returns(address synth) {
+    function deploySynth(address token) external onlyPOOLS returns (address synth) {
         require(getSynth[token] == address(0), "CreateErr");
         Synth newSynth;
-        newSynth = new Synth(token);  
+        newSynth = new Synth(token);
         synth = address(newSynth);
         _addSynth(token, synth);
         emit CreateSynth(token, synth);
     }
 
-    function mintSynth(address synth, address member, uint amount) external onlyPOOLS returns(bool) {
-         Synth(synth).mint(member, amount); 
+    function mintSynth(
+        address synth,
+        address member,
+        uint256 amount
+    ) external onlyPOOLS returns (bool) {
+        Synth(synth).mint(member, amount);
         return true;
     }
 
@@ -56,8 +59,7 @@ contract Factory {
 
     function _addSynth(address _token, address _synth) internal {
         getSynth[_token] = _synth;
-        arraySynths.push(_synth); 
+        arraySynths.push(_synth);
         isSynth[_synth] = true;
     }
-
 }
